move mk_norm() to add_book.match, since it is just a matching normalizer
